Updated file ensures build was successful

+++ update.sh
@@ -1,3 +1,4 @@
-docker-compose build
-docker-compose down
-docker-compose up -d
+if docker-compose build; then
+  docker-compose down
+  docker-compose up -d
+fi
